[#-] refactor: Move utils package to common for better structure
