+++ src/main/java/org/dnd/modutimer/common/exception/GlobalExceptionHandler.java
@@ -1,7 +1,7 @@
 package org.dnd.modutimer.common.exception;
 
 
-import org.dnd.modutimer.utils.ApiUtils;
+import org.dnd.modutimer.common.utils.ApiUtils;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.HttpMessageNotReadableException;
