+++ src/main/java/org/dnd/modutimer/common/utils/ApiUtils.java
@@ -1,4 +1,4 @@
-package org.dnd.modutimer.utils;
+package org.dnd.modutimer.common.utils;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
